[REM-103] refactor: update import paths and add main server entry point
- __main__.py: updated import path for main_server
- core/switchs/celery_beat.py: updated import path for StoreService and running_servers
- manager/runner.py: updated import path for StoreService and running_servers
- core/switchs/celery_worker.py: updated import path for StoreService and running_servers
- __init__.py: updated import paths for StoreService and HeadCrawjudManager
- core/main.py: added main server entry point implementation
