+++ crawjud/__main__.py
@@ -1,6 +1,6 @@
 """Main entry point for the server application."""
 
-from crawjud.main import main_server
+from crawjud.core.main import main_server
 
 if __name__ == "__main__":
     SystemExit(main_server())
