+++ crawjud/core/switchs/celery_worker.py
@@ -13,7 +13,7 @@ from quart import Quart
 from termcolor import colored
 from tqdm import tqdm
 
-from crawjud.config import StoreService, running_servers
+from crawjud.core.config import StoreService, running_servers
 from crawjud.core.watch import monitor_log
 from crawjud.utils import worker_name_generator
